Internship Task 1 has been completed
The Task given is to create a to-do web page using HTML, CSS, and Javascript. The page should incorporate the following features:

1) Ability to create a todo item
2) Capability to delete a todo item
3) Option to mark a todo item as complete
Please keep in mind that it is not necessary to save changes when the page reloads, meaning that data persistence is not required.

You are only allowed to use HTML, CSS, and Javascript for this task. The use of additional libraries or tools is not permitted.